chore: app store deployment services refactoring
