+++ client/cron/CdApplicationStatusUpdateHandler.go
@@ -12,8 +12,8 @@ import (
 	"github.com/devtron-labs/devtron/internal/sql/repository/pipelineConfig"
 	util2 "github.com/devtron-labs/devtron/internal/util"
 	"github.com/devtron-labs/devtron/pkg/app"
-	repository2 "github.com/devtron-labs/devtron/pkg/appStore/deployment/repository"
-	"github.com/devtron-labs/devtron/pkg/appStore/deployment/service"
+	repository2 "github.com/devtron-labs/devtron/pkg/appStore/installedApp/repository"
+	"github.com/devtron-labs/devtron/pkg/appStore/installedApp/service/EAMode"
 	"github.com/devtron-labs/devtron/pkg/pipeline"
 	"github.com/devtron-labs/devtron/util"
 	cron2 "github.com/devtron-labs/devtron/util/cron"
@@ -39,7 +39,7 @@ type CdApplicationStatusUpdateHandlerImpl struct {
 	cron                                 *cron.Cron
 	appService                           app.AppService
 	workflowDagExecutor                  pipeline.WorkflowDagExecutor
-	installedAppService                  service.InstalledAppService
+	installedAppService                  EAMode.InstalledAppDBService
 	CdHandler                            pipeline.CdHandler
 	AppStatusConfig                      *app.AppServiceConfig
 	pubsubClient                         *pubsub.PubSubClientServiceImpl
@@ -53,7 +53,7 @@ type CdApplicationStatusUpdateHandlerImpl struct {
 }
 
 func NewCdApplicationStatusUpdateHandlerImpl(logger *zap.SugaredLogger, appService app.AppService,
-	workflowDagExecutor pipeline.WorkflowDagExecutor, installedAppService service.InstalledAppService,
+	workflowDagExecutor pipeline.WorkflowDagExecutor, installedAppService EAMode.InstalledAppDBService,
 	CdHandler pipeline.CdHandler, AppStatusConfig *app.AppServiceConfig, pubsubClient *pubsub.PubSubClientServiceImpl,
 	pipelineStatusTimelineRepository pipelineConfig.PipelineStatusTimelineRepository,
 	eventClient client2.EventClient, appListingRepository repository.AppListingRepository,
