+++ pkg/appStore/installedApp/adapter/Adapter.go
