+++ pkg/appStore/installedApp/service/AppStoreDeploymentService_test.go
@@ -10,8 +10,8 @@ import (
 	"github.com/devtron-labs/devtron/internal/util"
 	appStoreBean "github.com/devtron-labs/devtron/pkg/appStore/bean"
 	repository6 "github.com/devtron-labs/devtron/pkg/appStore/chartGroup/repository"
-	repository3 "github.com/devtron-labs/devtron/pkg/appStore/deployment/repository"
 	appStoreDiscoverRepository "github.com/devtron-labs/devtron/pkg/appStore/discover/repository"
+	repository3 "github.com/devtron-labs/devtron/pkg/appStore/installedApp/repository"
 	repository5 "github.com/devtron-labs/devtron/pkg/auth/user/repository"
 	"github.com/devtron-labs/devtron/pkg/cluster"
 	repository2 "github.com/devtron-labs/devtron/pkg/cluster/repository"
