+++ pkg/appStore/installedApp/service/bean/bean.go
