+++ pkg/pipeline/WorkflowDagExecutor.go
@@ -187,18 +187,18 @@ type WorkflowDagExecutorImpl struct {
 	configMapHistoryRepository          repository3.ConfigMapHistoryRepository
 	helmAppService                      client2.HelmAppService
 	//TODO fix me next
-	helmAppClient                		gRPC.HelmAppClient //TODO refactoring: use helm app service instead
-	environmentConfigRepository  		chartConfig.EnvConfigOverrideRepository
-	mergeUtil                    		*util.MergeUtil
-	acdClient                    		application2.ServiceClient
-	argoClientWrapperService     		argocdServer.ArgoClientWrapperService
-	customTagService             		CustomTagService
-	ACDConfig                    		*argocdServer.ACDConfig
-	deployedAppMetricsService    		deployedAppMetrics.DeployedAppMetricsService
-	chartRefService              		chartRef.ChartRefService
-	gitOpsConfigReadService             config.GitOpsConfigReadService
-	gitOperationService                 git.GitOperationService
-	imageDigestPolicyService            imageDigestPolicy.ImageDigestPolicyService
+	helmAppClient               gRPC.HelmAppClient //TODO refactoring: use helm app service instead
+	environmentConfigRepository chartConfig.EnvConfigOverrideRepository
+	mergeUtil                   *util.MergeUtil
+	acdClient                   application2.ServiceClient
+	argoClientWrapperService    argocdServer.ArgoClientWrapperService
+	customTagService            CustomTagService
+	ACDConfig                   *argocdServer.ACDConfig
+	deployedAppMetricsService   deployedAppMetrics.DeployedAppMetricsService
+	chartRefService             chartRef.ChartRefService
+	gitOpsConfigReadService     config.GitOpsConfigReadService
+	gitOperationService         git.GitOperationService
+	imageDigestPolicyService    imageDigestPolicy.ImageDigestPolicyService
 }
 
 const kedaAutoscaling = "kedaAutoscaling"
@@ -2489,7 +2489,7 @@ func (impl *WorkflowDagExecutorImpl) ManualCdTrigger(triggerContext TriggerConte
 		overrideRequest.CdWorkflowId = cdWorkflowId
 		// creating cd pipeline status timeline for deployment initialisation
 		timeline := impl.pipelineStatusTimelineService.GetTimelineDbObjectByTimelineStatusAndTimelineDescription(savedWfr.Id, 0, pipelineConfig.TIMELINE_STATUS_DEPLOYMENT_INITIATED, pipelineConfig.TIMELINE_DESCRIPTION_DEPLOYMENT_INITIATED, overrideRequest.UserId, time.Now())
-		_, span = otel.Tracer("orchestrator").Start(ctx, "cdPipelineStatusTimelineRepo.SaveTimelineForACDHelmApps")
+		_, span = otel.Tracer("orchestrator").Start(ctx, "cdPipelineStatusTimelineRepo.SaveTimelineForHelmApps")
 		err = impl.pipelineStatusTimelineService.SaveTimeline(timeline, nil, false)
 
 		span.End()
